bookstore-microservice/staff-service/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
django-cors-headers>=4.3.0
requests>=2.31.0
